question 12, removed Inputs and Outputs to use Service instead
