+++ src/app/pet/pet-detail/pet-detail.component.ts
@@ -1,20 +1,18 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { IPet } from '../model/pet';
+import { PetService } from '../pet.service';
 
 @Component({
   selector: 'app-pet-detail',
   templateUrl: './pet-detail.component.html',
   styleUrls: ['./pet-detail.component.scss']
 })
-export class PetDetailComponent implements OnInit {
-
-  @Input() pet: IPet|undefined|null;
+export class PetDetailComponent {
   
-  constructor() {}
+  constructor(private petService: PetService) {}
 
-  ngOnInit(): void {
-
+  get pet(): IPet|undefined|null {
+    return this.petService.selectedPet;
   }
-  
 
 }
