+++ src/app/pet/pet.component.html
@@ -1,9 +1,4 @@
 <div class="pet">
-    <app-pet-list
-     [petList]="pets"
-     (selectPet)="onSelectPet($event)"
-     ></app-pet-list>
-    <app-pet-detail
-        [pet]="selectedPet"
-    ></app-pet-detail>
+    <app-pet-list></app-pet-list>
+    <app-pet-detail></app-pet-detail>
 </div>
